pyproject.toml
[project]
name = "local-codereview"
version = "0.1.0"
description = "A local CLI tool for Git-based code reviews with LLM integration."
requires-python = ">=3.9"
dependencies = [
    "questionary",  # For interactive CLI prompts
    "python-dotenv",# For loading .env files
    "google-generativeai", # For Google Gemini API
]

[project.scripts]
codereview = "codereview_tool.cli:main" # Entry point for the CLI

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
